# cmd/server/main.go
package main

import (
	"context"
	"encoding/json"
	"errors"
	"log"
	"net"
	"net/http"
	"os"
	"os/signal"
	"strconv"
	"strings"
	"syscall"
	"time"

	"github.com/ekucher/bsystem-integration-core/internal/ai"
	"github.com/ekucher/bsystem-integration-core/internal/authz"
	"github.com/ekucher/bsystem-integration-core/internal/platformdb"
	"github.com/ekucher/bsystem-integration-core/internal/search"
	"github.com/nats-io/nats.go"
)

type healthResponse struct {
	Status    string            `json:"status"`
	Service   string            `json:"service"`
	Version   string            `json:"version"`
	Timestamp string            `json:"timestamp"`
	Checks    map[string]string `json:"checks"`
}

type userInfo struct {
	Sub               string   `json:"sub"`
	Email             string   `json:"email"`
	Name              string   `json:"name"`
	PreferredUsername string   `json:"preferred_username"`
	Groups            []string `json:"groups"`
}

type meResponse struct {
	ID          string   `json:"id"`
	Subject     string   `json:"subject"`
	Email       string   `json:"email"`
	Name        string   `json:"name"`
	Username    string   `json:"username"`
	Groups      []string `json:"groups"`
	Roles       []string `json:"roles"`
	Permissions []string `json:"permissions"`
	Modules     []string `json:"modules"`
}

type globalIDRequest struct {
	EntityType string         `json:"entity_type"`
	Source     string         `json:"source"`
	SourceID   string         `json:"source_id"`
	TenantID   string         `json:"tenant_id,omitempty"`
	Metadata   map[string]any `json:"metadata,omitempty"`
}

type contextKey string

const (
	// userContextKey holds the resolved meResponse for a human caller.
	userContextKey contextKey = "user"
	// principalContextKey holds the authorization principal, whether the
	// caller is a human or a service identity.
	principalContextKey contextKey = "principal"
	requestIDContextKey contextKey = "request-id"
)

type app struct {
	db    *platformdb.DB
	nc    *nats.Conn
	authz *authz.Evaluator
	// aiProvider is the language model the gateway calls. It is always
	// non-nil: the fake provider is the default, so the gateway's
	// authorization, classification and audit behaviour is exercised in every
	// deployment rather than only where a model is configured.
	aiProvider ai.Provider
	// searchProvider is the search engine. It is always non-nil: the
	// in-memory provider is the default, so search answers honestly with an
	// empty index rather than failing as unconfigured.
	searchProvider search.Provider
}

// accessFrom returns the resolved access of the human caller.
func accessFrom(ctx context.Context) meResponse {
	access, _ := ctx.Value(userContextKey).(meResponse)
	return access
}

// principalFrom returns the authorization principal of the caller.
func principalFrom(ctx context.Context) authz.Principal {
	principal, _ := ctx.Value(principalContextKey).(authz.Principal)
	return principal
}

// writeDenied renders an authorization denial. The decision's reason names the
// missing permission and nothing else, so a denial never reveals whether the
// resource exists or who owns it.
func writeDenied(w http.ResponseWriter, decision authz.Decision) {
	body := map[string]string{"error": decision.Reason}
	if decision.Code != "" {
		body["code"] = decision.Code
	}
	writeJSON(w, http.StatusForbidden, body)
}

// authorize enforces a route's declared permission across the whole platform,
// before the handler runs.
//
// Routes that address a single resource declare a scope type instead, and
// evaluate in the handler once they know which resource was addressed; see
// authorizeResource.
func (a *app) authorize(permission string, next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		decision, err := a.authz.Evaluate(r.Context(), principalFrom(r.Context()), permission, authz.Global())
		if err != nil {
			logger.ErrorContext(r.Context(), "authorization evaluation failed", "error", err.Error())
			writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "authorization store unavailable"})
			return
		}
		if !decision.Allowed {
			writeDenied(w, decision)
			return
		}
		next.ServeHTTP(w, r)
	})
}

// authorizeResource evaluates a route's permission against one resolved
// resource. It reports whether the caller may proceed, having already written
// the denial when they may not.
func (a *app) authorizeResource(w http.ResponseWriter, r *http.Request, permission, scopeType, scopeID string) bool {
	decision, err := a.authz.Evaluate(r.Context(), principalFrom(r.Context()), permission, authz.Resource(scopeType, scopeID))
	if err != nil {
		logger.ErrorContext(r.Context(), "authorization evaluation failed", "error", err.Error())
		writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "authorization store unavailable"})
		return false
	}
	if !decision.Allowed {
		writeDenied(w, decision)
		return false
	}
	return true
}

func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(v)
}

func requestID(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		id := r.Header.Get("X-Request-ID")
		if id == "" {
			id = time.Now().UTC().Format("20060102T150405.000000000Z07:00")
		}
		w.Header().Set("X-Request-ID", id)
		ctx := context.WithValue(r.Context(), requestIDContextKey, id)
		// The store layer queues durable events inside the transactions it
		// runs, and an event nobody can trace back to a request is half a
		// trail. Set once here rather than threaded through every call.
		ctx = platformdb.WithCorrelationID(ctx, id)
		next.ServeHTTP(w, r.WithContext(ctx))
	})
}

func fetchUserInfo(ctx context.Context, token string) (userInfo, error) {
	endpoint := os.Getenv("AUTHENTIK_USERINFO_URL")
	if endpoint == "" {
		return userInfo{}, errors.New("AUTHENTIK_USERINFO_URL is not configured")
	}
	req, err := http.NewRequestWithContext(ctx, http.MethodGet, endpoint, nil)
	if err != nil {
		return userInfo{}, err
	}
	req.Header.Set("Authorization", "Bearer "+token)
	req.Header.Set("Accept", "application/json")
	client := &http.Client{Timeout: 5 * time.Second}
	resp, err := client.Do(req)
	if err != nil {
		return userInfo{}, err
	}
	defer resp.Body.Close()
	if resp.StatusCode != http.StatusOK {
		return userInfo{}, errors.New("token rejected by authentik")
	}
	var info userInfo
	if err := json.NewDecoder(resp.Body).Decode(&info); err != nil {
		return userInfo{}, err
	}
	if info.Sub == "" {
		return userInfo{}, errors.New("userinfo response has no subject")
	}
	return info, nil
}

func (a *app) authenticate(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		header := r.Header.Get("Authorization")
		if !strings.HasPrefix(header, "Bearer ") {
			writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "missing bearer token"})
			return
		}
		info, err := fetchUserInfo(r.Context(), strings.TrimPrefix(header, "Bearer "))
		if err != nil {
			logger.WarnContext(r.Context(), "authentication failed", "error", err.Error())
			writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "invalid or expired token"})
			return
		}
		username := info.PreferredUsername
		if username == "" {
			username = info.Email
		}
		// The "created" flag used to decide whether to publish identity.created
		// from here. The allocation queues that event in its own transaction
		// now, so the caller no longer needs to know.
		globalUserID, _, err := a.db.EnsureIdentity(r.Context(), platformdb.Identity{
			Subject: info.Sub, Email: info.Email, DisplayName: info.Name, Username: username, Groups: unique(info.Groups),
		})
		if err != nil {
			logger.ErrorContext(r.Context(), "identity persistence failed", "error", err.Error())
			writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "identity persistence unavailable"})
			return
		}
		// Access is resolved once per request rather than per handler, so a
		// single request cannot observe two different authorization states.
		access, err := a.resolveAccess(r.Context(), info, globalUserID)
		if err != nil {
			logger.ErrorContext(r.Context(), "RBAC resolution failed", "error", err.Error())
			writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "RBAC store unavailable"})
			return
		}
		ctx := context.WithValue(r.Context(), userContextKey, access)
		ctx = context.WithValue(ctx, principalContextKey, authz.Principal{
			ID: access.ID, Kind: authz.KindUser, Roles: access.Roles, Permissions: access.Permissions,
		})
		next.ServeHTTP(w, r.WithContext(ctx))
	})
}

func unique(values []string) []string {
	seen := map[string]bool{}
	result := make([]string, 0, len(values))
	for _, value := range values {
		if value != "" && !seen[value] {
			seen[value] = true
			result = append(result, value)
		}
	}
	return result
}

func (a *app) resolveAccess(ctx context.Context, info userInfo, globalUserID string) (meResponse, error) {
	profile, err := a.db.ResolveAccess(ctx, unique(info.Groups), "human")
	if err != nil {
		return meResponse{}, err
	}
	username := info.PreferredUsername
	if username == "" {
		username = info.Email
	}
	return meResponse{
		ID:          globalUserID,
		Subject:     info.Sub,
		Email:       info.Email,
		Name:        info.Name,
		Username:    username,
		Groups:      unique(info.Groups),
		Roles:       profile.Roles,
		Permissions: profile.Permissions,
		Modules:     profile.Modules,
	}, nil
}

func requestIDFrom(ctx context.Context) string {
	id, _ := ctx.Value(requestIDContextKey).(string)
	return id
}

func sourceIP(r *http.Request) string {
	if forwarded := strings.TrimSpace(strings.Split(r.Header.Get("X-Forwarded-For"), ",")[0]); forwarded != "" {
		return forwarded
	}
	host, _, err := net.SplitHostPort(r.RemoteAddr)
	if err == nil {
		return host
	}
	return r.RemoteAddr
}

func (a *app) audit(r *http.Request, access meResponse, action, resourceType, resourceID string, metadata map[string]any) {
	err := a.db.InsertAudit(r.Context(), platformdb.AuditEvent{Subject: access.Subject, GlobalUserID: access.ID, Action: action, ResourceType: resourceType, ResourceID: resourceID, RequestID: requestIDFrom(r.Context()), SourceIP: sourceIP(r), Metadata: metadata})
	if err != nil {
		auditWrites.Inc(action, "failed")
		// The request is not failed here. The action it describes has already
		// happened and telling the caller otherwise would be a lie in the
		// other direction — but an action that happened without a record is a
		// hole in the governance layer, so it is counted rather than only
		// logged.
		logger.ErrorContext(r.Context(), "audit write failed", "action", action, "error", err.Error())
		return
	}
	auditWrites.Inc(action, "written")
}

// publish sends a platform event, recording the outcome.
//
// The outcome is counted rather than only logged, because "no events" and
// "every event failed to publish" look identical on a dashboard that only
// counts successes.
func (a *app) publish(subject string, payload any) {
	if a.nc == nil {
		eventsPublished.Inc(subject, "unavailable")
		return
	}
	body, err := json.Marshal(payload)
	if err != nil {
		eventsPublished.Inc(subject, "encode_failed")
		logger.Error("event could not be encoded", "subject", subject, "error", err.Error())
		return
	}
	if err := a.nc.Publish(subject, body); err != nil {
		eventsPublished.Inc(subject, "failed")
		logger.Error("event publication failed", "subject", subject, "error", err.Error())
		return
	}
	eventsPublished.Inc(subject, "published")
}

func (a *app) health(w http.ResponseWriter, r *http.Request) {
	checks := map[string]string{"database": "ok", "nats": "ok"}
	status := "ok"
	ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
	defer cancel()
	if err := a.db.Ping(ctx); err != nil {
		checks["database"] = "error"
		status = "degraded"
	}
	if a.nc == nil || !a.nc.IsConnected() {
		checks["nats"] = "degraded"
		status = "degraded"
	}
	code := http.StatusOK
	if checks["database"] == "error" {
		code = http.StatusServiceUnavailable
	}
	writeJSON(w, code, healthResponse{Status: status, Service: "bsystem-integration-core", Version: "0.5.0", Timestamp: time.Now().UTC().Format(time.RFC3339), Checks: checks})
}

func (a *app) me(w http.ResponseWriter, r *http.Request) {
	writeJSON(w, http.StatusOK, accessFrom(r.Context()))
}

func (a *app) modules(w http.ResponseWriter, r *http.Request) {
	access := accessFrom(r.Context())
	allowed := map[string]bool{}
	for _, id := range access.Modules {
		allowed[id] = true
	}
	all, err := a.db.ListModules(r.Context())
	if err != nil {
		writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "module registry unavailable"})
		return
	}
	result := []platformdb.Module{}
	for _, item := range all {
		if allowed[item.ID] {
			result = append(result, item)
		}
	}
	writeJSON(w, http.StatusOK, result)
}

func (a *app) createGlobalID(w http.ResponseWriter, r *http.Request) {
	access := accessFrom(r.Context())
	var input globalIDRequest
	if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, 1<<20)).Decode(&input); err != nil {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid JSON body"})
		return
	}
	input.EntityType = strings.TrimSpace(input.EntityType)
	input.Source = strings.TrimSpace(input.Source)
	input.SourceID = strings.TrimSpace(input.SourceID)
	if input.EntityType == "" || input.Source == "" || input.SourceID == "" {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "entity_type, source and source_id are required"})
		return
	}
	entity, err := a.db.CreateGlobalEntity(r.Context(), input.EntityType, input.Source, input.SourceID, strings.TrimSpace(input.TenantID), input.Metadata)
	if err != nil {
		// Only one failure here is the caller's, and it is the only one whose
		// text is safe to return. Everything else is the platform's, and
		// err.Error() on a database failure is a raw SQL error: it names the
		// table, the column tuple and the constraint, which is internal schema
		// topology handed to whoever asked.
		//
		// Answering 400 for those was wrong twice over — it told a caller
		// their valid input was invalid, and 400 is the status nobody retries.
		if errors.Is(err, platformdb.ErrUnsupportedEntityType) {
			writeJSON(w, http.StatusBadRequest, map[string]string{
				"error": err.Error(),
				"code":  "unsupported_entity_type",
			})
			return
		}
		logger.ErrorContext(r.Context(), "global id allocation failed", "entity_type", input.EntityType, "source", input.Source, "error", err.Error())
		writeJSON(w, http.StatusInternalServerError, map[string]string{
			"error":      "global id could not be allocated",
			"code":       "global_id_allocation_failed",
			"request_id": requestIDFrom(r.Context()),
		})
		return
	}
	a.audit(r, access, "global_id.created", input.EntityType, entity.GlobalID, map[string]any{"source": input.Source, "source_id": input.SourceID})
	// The event is queued by the allocation itself, in its transaction, and
	// only when an identifier was actually minted. See
	// internal/platformdb/outbox.go.
	writeJSON(w, http.StatusCreated, entity)
}

func (a *app) resolveGlobalID(w http.ResponseWriter, r *http.Request) {
	access := accessFrom(r.Context())
	id := strings.TrimSpace(r.PathValue("id"))
	if id == "" {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "Global ID is required"})
		return
	}
	entity, err := a.db.ResolveGlobalEntity(r.Context(), id)
	if err != nil {
		writeJSON(w, http.StatusNotFound, map[string]string{"error": "Global ID not found"})
		return
	}
	a.audit(r, access, "global_id.read", entity.EntityType, entity.GlobalID, nil)
	writeJSON(w, http.StatusOK, entity)
}

func (a *app) auditEvents(w http.ResponseWriter, r *http.Request) {
	limit, _ := strconv.Atoi(r.URL.Query().Get("limit"))
	events, err := a.db.ListAudit(r.Context(), limit)
	if err != nil {
		writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "audit store unavailable"})
		return
	}
	writeJSON(w, http.StatusOK, events)
}

func main() {
	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
	defer cancel()
	db, err := platformdb.Open(ctx, os.Getenv("DATABASE_URL"))
	if err != nil {
		log.Fatalf("database initialization failed: %v", err)
	}
	defer db.Close()

	var nc *nats.Conn
	if natsURL := os.Getenv("NATS_URL"); natsURL != "" {
		// RetryOnFailedConnect is what makes a broker that is down at startup
		// a delay rather than a permanent condition. Without it a Core that
		// booted a second before NATS came up held a nil connection for the
		// rest of its life: every event it produced was counted "unavailable"
		// and dropped, /readyz reported nats degraded forever, and only a
		// restart fixed it. The durable outbox would have queued behind a
		// broker this process had decided did not exist.
		nc, err = nats.Connect(natsURL,
			nats.Name("bsystem-integration-core"),
			nats.Timeout(5*time.Second),
			nats.RetryOnFailedConnect(true),
			nats.MaxReconnects(-1))
		if err != nil {
			logger.Warn("NATS unavailable at startup", "error", err.Error())
			nc = nil
		} else {
			defer nc.Close()
		}
	}
	a := &app{db: db, nc: nc, authz: authz.New(db, authz.DefaultConfinedRoles()), searchProvider: searchProvider(), aiProvider: aiProviderFromEnv()}
	a.registerPlatformMetrics()
	a.registerBuildMetrics()

	// The durable half of event delivery. It runs whether or not NATS was
	// reachable at startup: events keep being queued by the transactions that
	// cause them, and this drains them when a broker is there. A core that
	// started without one therefore still delivers, which the request-path
	// publish it replaced could not do.
	outboxCtx, stopOutbox := context.WithCancel(context.Background())
	defer stopOutbox()
	if nc != nil {
		go a.runOutbox(outboxCtx, newJetStreamPublisher(nc), outboxInterval())
	} else {
		logger.Warn("durable events will be queued but not delivered", "reason", "NATS_URL is not configured or was unreachable at startup")
	}

	// Said once, at the moment somebody is watching a deployment. A metric is
	// how this stays visible afterwards; a log line is how it gets noticed at
	// all. This does not refuse to start: the schema is applied and the
	// service works, and turning a bookkeeping divergence into an outage
	// would make the honest thing to do the expensive one.
	if state, err := db.SchemaLevel(ctx); err == nil && len(state.Drifted) > 0 {
		logger.Error("migration files no longer match what this database applied",
			"migrations", strings.Join(state.Drifted, ","),
			"level", state.Level,
			"detail", "this database and one migrated from the current source report the same schema level and hold different schemas")
	}

	addr := os.Getenv("HTTP_ADDR")
	if addr == "" {
		addr = ":8080"
	}
	// WriteTimeout must exceed the longest a handler can legitimately take,
	// or the server cuts a response the handler is still producing — and the
	// caller sees a truncated body rather than the timeout that caused it.
	// The AI gateway is the longest, so the bound is derived from it rather
	// than written as a constant somebody has to remember to keep in step.
	writeTimeout := durationEnv("HTTP_WRITE_TIMEOUT", 60*time.Second)
	if minimum := aiTimeout() + 10*time.Second; writeTimeout < minimum {
		logger.Warn("raising the write timeout above the AI provider bound",
			"configured", writeTimeout.String(), "using", minimum.String())
		writeTimeout = minimum
	}

	server := &http.Server{
		Addr:    addr,
		Handler: a.handler(),
		// ReadHeaderTimeout is the one that matters for a slowloris: it
		// bounds how long a connection may spend sending nothing useful.
		ReadHeaderTimeout: durationEnv("HTTP_READ_HEADER_TIMEOUT", 5*time.Second),
		ReadTimeout:       durationEnv("HTTP_READ_TIMEOUT", 15*time.Second),
		WriteTimeout:      writeTimeout,
		IdleTimeout:       durationEnv("HTTP_IDLE_TIMEOUT", 60*time.Second),
	}

	// Shut down on SIGTERM rather than dying on it.
	//
	// A container runtime sends SIGTERM and then waits before SIGKILL. A
	// process that exits immediately drops every request in flight, which
	// during a rolling deploy means a burst of failures for users who did
	// nothing but arrive at the wrong moment — and which looks like an
	// intermittent platform fault rather than a deployment.
	shutdown := make(chan os.Signal, 1)
	signal.Notify(shutdown, syscall.SIGINT, syscall.SIGTERM)

	serverErrors := make(chan error, 1)
	go func() {
		logger.Info("bsystem-integration-core listening", "version", "0.6.0", "addr", addr,
			"write_timeout", writeTimeout.String())
		serverErrors <- server.ListenAndServe()
	}()

	if err := serveUntilSignal(server, serverErrors, shutdown,
		durationEnv("HTTP_SHUTDOWN_GRACE", 10*time.Second)); err != nil {
		log.Fatalf("server failed: %v", err)
	}
}

// serveUntilSignal waits for the server to fail or for a shutdown signal, and
// on a signal drains the requests already in flight within grace.
//
// It is a function rather than the tail of main so that the draining can be
// tested. The property it exists for — a request that arrived before SIGTERM
// finishes rather than being dropped — is invisible on a healthy platform and
// appears only during a rolling deploy, as a burst of failures for users who
// did nothing but arrive at the wrong moment. That reads as an intermittent
// platform fault rather than as a deployment, which is what makes it expensive
// to diagnose and worth pinning.
func serveUntilSignal(server *http.Server, serverErrors <-chan error, signals <-chan os.Signal, grace time.Duration) error {
	select {
	case err := <-serverErrors:
		if err != nil && !errors.Is(err, http.ErrServerClosed) {
			return err
		}
		return nil
	case sig := <-signals:
		logger.Info("shutting down", "signal", sig.String())
		// The grace period must be shorter than the runtime's own kill
		// delay, or the runtime wins the race and the graceful path never
		// completes. Ten seconds is comfortably inside Docker's default
		// thirty and Kubernetes' default thirty.
		ctx, cancel := context.WithTimeout(context.Background(), grace)
		defer cancel()
		if err := server.Shutdown(ctx); err != nil {
			// A request that outlived the grace period is closed. Saying so
			// is the point: a silent close looks like a network fault to
			// whoever was holding the connection.
			logger.Warn("shutdown deadline reached with requests still in flight", "error", err.Error())
			_ = server.Close()
		}
		logger.Info("stopped")
		return nil
	}
}