# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BSYSTEM Integration Core API
  version: 0.8.0
  summary: Authorization-aware normalized API for the BSYSTEM Platform.
  description: |
    The Integration Core exposes a normalized view of the systems BSYSTEM
    integrates with. Source systems stay authoritative; this API never
    re-exposes a raw upstream payload.

    Three surfaces are served:

    - `/api/v1/*` — the human API, authenticated with a user access token.
    - `/api/service/v1/*` — the machine API, authenticated with a service
      identity in the `BSYSTEM-Services` group.
    - `/health`, `/readyz`, `/metrics` — operational endpoints, unauthenticated
      by design, which disclose no business data.

    Authorization is enforced in the backend for every request and denies by
    default. An unknown or unmapped resource results in denial, never in
    broader access.

    Every response carries `X-Request-ID`, echoing the caller's value or a
    generated one, and the same identifier reaches the audit trail and
    published events.

    This document is contract-tested against the server's own route inventory:
    an endpoint cannot be added or removed without the documentation changing
    with it.
  contact:
    name: BSYSTEM Platform
    url: https://github.com/ekucher/bsystem-integration-core
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: The Integration Core, behind the platform reverse proxy.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/modules:
    get:
      tags: [Identity]
      operationId: listModules
      summary: List the modules the authenticated user may see
      responses:
        '200':
          description: The modules the caller's roles grant, filtered server-side.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Module'}
              example:
                - {id: projects, name: Projects, description: Project and task management, status: available}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
  /api/v1/global-ids:
    post:
      tags: [Global IDs]
      operationId: createGlobalID
      summary: Allocate or resolve a Global ID mapping
      description: |
        Administrator only. The operation is idempotent: mapping the same
        `(source, entity_type, source_id)` again returns the identifier
        already allocated. Global IDs are immutable once issued.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GlobalIDRequest'}
            example: {entity_type: client, source: espocrm, source_id: acc-northwind, metadata: {name: Northwind Trading}}
      responses:
        '201':
          description: The mapping, newly allocated or already existing.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GlobalEntity'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/global-ids/{id}:
    get:
      tags: [Global IDs]
      operationId: resolveGlobalID
      summary: Resolve a Global ID to its upstream mapping
      description: Administrator only. Reads are audited.
      parameters:
        - name: id
          in: path
          required: true
          description: The Global ID to resolve.
          schema: {type: string, pattern: '^[A-Z]+-[0-9]+$'}
          example: CL-000001
      responses:
        '200':
          description: The mapping.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GlobalEntity'}
              example: {global_id: CL-000001, entity_type: client, source: espocrm, source_id: acc-northwind, created_at: '2026-01-06T11:30:00Z'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OAuth2 access token
      description: A human access token issued by authentik.
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: OAuth2 access token
      description: |
        A service identity access token issued by authentik. The identity must
        be a member of `BSYSTEM-Services`.
...
  responses:
    BadRequest:
      description: |
        The request is malformed, missing a required field, or carries a
        pagination cursor the platform did not issue.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: invalid pagination cursor, code: invalid_cursor}
    Unauthorized:
      description: The access token is missing, invalid or expired.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: invalid or expired token}
    Forbidden:
      description: |
        The caller is authenticated but is not authorized.

        `permission_required` means the caller does not hold the permission.
        `scope_required` means the caller holds it but may only exercise it
        inside a scope they have been granted, and no grant matches.

        The body names the missing permission and nothing else: it never
        discloses whether the resource exists or who owns it.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: crm.client.read permission required, code: permission_required}
...
    NotFound:
      description: |
        The resource does not exist, or the caller may not learn that it does.
        The two are deliberately indistinguishable.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: resource not found, code: not_found}
...
    DependencyUnavailable:
      description: |
        A platform dependency is unavailable, or an integration this endpoint
        needs is not configured.

        On an adapter-backed endpoint this is the answer a deployment gets when
        it deliberately leaves a source system out, which is a supported
        configuration rather than a fault. Those answers carry
        `adapter_not_configured` or `capability_unsupported` and name the
        adapter in `source`, so a caller can say which integration is missing
        instead of reporting an unexplained failure. A dependency of the
        platform itself, such as the database, carries no code.

        This is never rendered as an empty collection. An empty result would
        tell a caller the platform knows of no records, when the truth is that
        nobody has told it where to look.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            adapterNotConfigured:
              value: {error: EspoCRM adapter not configured, code: adapter_not_configured, source: espocrm}
            platformDependency:
              value: {error: RBAC store unavailable}
  schemas:
...
    Module:
      type: object
      required: [id, name, description, status]
      properties:
        id: {type: string}
        name: {type: string}
        description: {type: string}
        status: {type: string}
...
    GlobalIDRequest:
      type: object
      required: [entity_type, source, source_id]
      properties:
        entity_type:
          type: string
          description: The platform entity type, which selects the identifier prefix.
          example: client
        source: {type: string, description: The authoritative system., example: espocrm}
        source_id: {type: string, description: The identifier within that system.}
        tenant_id: {type: string, description: The owning tenant, when ownership is known.}
        metadata:
          type: object
          description: Integration metadata. Must not carry confidential upstream content.
          additionalProperties: true
    GlobalEntity:
      type: object
      required: [global_id, entity_type, source, source_id, created_at]
      properties:
        global_id: {type: string, pattern: '^[A-Z]+-[0-9]+$'}
        entity_type: {type: string}
        source: {type: string}
        source_id: {type: string}
        tenant_id: {type: string}
        metadata: {type: object, additionalProperties: true}
        created_at: {type: string, format: date-time}
...